refactor: move describe_table_api to control
API should be called in control
