+++ src/app.rs
@@ -16,10 +16,7 @@
 
 use ::serde::{Deserialize, Serialize};
 use log::{debug, error, info};
-use rusoto_dynamodb::{
-    AttributeDefinition, DescribeTableInput, DynamoDb, DynamoDbClient, KeySchemaElement,
-    TableDescription,
-};
+use rusoto_dynamodb::{AttributeDefinition, KeySchemaElement, TableDescription};
 use rusoto_signature::Region;
 use serde_yaml::Error as SerdeYAMLError;
 use std::{
@@ -442,7 +439,7 @@ pub async fn use_table(
             debug!("describing the table: {}", tbl);
             let region = cx.effective_region();
             let tbl = tbl.clone();
-            let desc: TableDescription = describe_table_api(&region, tbl.clone()).await;
+            let desc: TableDescription = control::describe_table_api(&region, tbl.clone()).await;
             save_using_target(cx, desc)?;
             println!("Now you're using the table '{}' ({}).", tbl, &region.name());
         },
@@ -555,7 +552,7 @@ pub async fn table_schema(cx: &Context) -> TableSchema {
         // It's possible that users pass --table without calling `dy use` for any table. Thus collect all data from DescribeTable results.
         Some(table_name) => {
             // TODO: reduce # of DescribeTable API calls. table_schema function is called every time you do something.
-            let desc: TableDescription = describe_table_api(
+            let desc: TableDescription = control::describe_table_api(
                 &cx.effective_region(),
                 table_name, /* should be equal to 'cx.effective_table_name()' */
             )
@@ -622,26 +619,6 @@ pub fn index_schemas(desc: &TableDescription) -> Option<Vec<IndexSchema>> {
         None
     } else {
         Some(indexes)
-    }
-}
-
-/// Originally intended to be called by describe_table function, which is called from `$ dy desc`,
-/// however it turned out that DescribeTable API result is useful in various logic, separated API into this standalone function.
-pub async fn describe_table_api(region: &Region, table_name: String) -> TableDescription {
-    let ddb = DynamoDbClient::new(region.clone());
-    let req: DescribeTableInput = DescribeTableInput { table_name };
-
-    match ddb.describe_table(req).await {
-        Err(e) => {
-            debug!("DescribeTable API call got an error -- {:#?}", e);
-            error!("{}", e.to_string());
-            std::process::exit(1);
-        }
-        Ok(res) => {
-            let desc: TableDescription = res.table.expect("This message should not be shown.");
-            debug!("Received DescribeTable Result: {:?}\n", desc);
-            desc
-        }
     }
 }
